+++ elipsoid/eig/eig3.h
@@ -6,6 +6,6 @@
 
 /* Symmetric matrix A => eigenvectors in columns of V, corresponding
    eigenvalues in d. */
-void eigen_decomposition(double A[3][3], double V[3][3], double d[3]);
+void eigen_decomposition(float A[3][3], float V[3][3], float d[3]);
 
 #endif
